refactor: 멀티모듈 구조 전환
