+++ backend/.gitignore
@@ -35,3 +35,7 @@ out/
 
 ### VS Code ###
 .vscode/
+
+### Environment ###
+*.env
+.env.*
