+++ backend/flowmeet-api/src/test/java/kr/flowmeet/api/FlowmeetApiApplicationTests.java
@@ -1,10 +1,10 @@
-package capstone.backend;
+package kr.flowmeet.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BackendApplicationTests {
+class FlowmeetApiApplicationTests {
 
     @Test
     void contextLoads() {
