+++ backend/flowmeet-common/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+    implementation project(':flowmeet-global-utils')
+}
