+++ backend/flowmeet-global-utils/build.gradle
@@ -1,0 +1,2 @@
+dependencies {
+}
