+++ backend/settings.gradle
@@ -1,1 +1,4 @@
-rootProject.name = 'backend'
+rootProject.name = 'flowmeet'
+
+include 'flowmeet-api', 'flowmeet-auth', 'flowmeet-common',
+        'flowmeet-domain', 'flowmeet-external', 'flowmeet-global-utils'
