--- backend/src/main/resources/application.yaml
@@ -1,3 +1,0 @@
-spring:
-  application:
-    name: backend
